chore: rename the checkbox_icon.js file
