+++ src/flyout_checkbox_icon.js
